requirements.txt
selenium==4.23.1
pytest==7.4.0
webdriver-manager
PyYaml
allure-pytest